ML_Final_Project/requirements.txt
Flask==2.0.1
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2
seaborn==0.11.1
tensorflow==2.5.0
scikit-learn==0.24.2
gunicorn==20.1.0
flask-cors==3.0.10
google-cloud-storage==1.42.3